ethernet: change MTU type to size_t
Modify the type to get this working on 64-bit system.

In phosphor-dbus-interfaces, the type is `size`:
https://github.com/openbmc/phosphor-dbus-interfaces/blob/8a2674b/yaml/xyz/openbmc_project/Network/EthernetInterface.interface.yaml#L26

Tested:
```
Before:
 curl localhost/redfish/v1/Managers/bmc/EthernetInterfaces/eth0 -s | grep MTU
  "MTUSize": 0,
After:
 curl localhost/redfish/v1/Managers/bmc/EthernetInterfaces/eth0 -s | grep MTU
  "MTUSize": 1500,
```

+++ redfish-core/lib/ethernet.hpp
@@ -254,8 +254,8 @@ inline bool extractEthernetInterfaceData(
                         }
                         else if (propertyPair.first == "MTU")
                         {
-                            const uint32_t* mtuSize =
-                                std::get_if<uint32_t>(&propertyPair.second);
+                            const size_t* mtuSize =
+                                std::get_if<size_t>(&propertyPair.second);
                             if (mtuSize != nullptr)
                             {
                                 ethData.mtuSize = *mtuSize;
